[FIX] mail: chat window should remain folded on new message
If the chat window was initially folded and when a new message is received
the chat window automatically unfolds. it is annoying having a window that
pops up each time a new message is coming while having folded it once.
So, It should remain folded when a new message is received if the chat
window is already folded.

Task : 2522141

closes odoo/odoo#71258

X-original-commit: ab00630b46cbdc2b961c6b52be281427535f4820
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>