[FIX] payment_ogone: accept tokenizing cards with missing number
In Ogone's backend, it is possible to remove `CARD.CARDNUMBER` from
Flexcheckout's feedback data (Configuration -> Technical information ->
Transaction feedback -> Alias gateway and Tokenization -> Dynamic
parameters). Since this is the configuration that was suggested to Odoo
customers migrating from a previous version, we must be able to process
feedback data when the card number is missing.

This commit creates an anonymous card number ('XXXX...') if one is not
provided by Ogone.

task-2494916

closes odoo/odoo#70896

X-original-commit: d1699f44bf3e57199387959af47a871224f20cbc
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>