[FIX] base: confusion between werkzeug.urls and urls local
As part of the Python 3 compatibility effort,
818ffdf600bb797ab66fae2628e8780804bf2181 moved all uses of urllib(2)
to werkzeug & requests as the packages were renamed and reorganised
between P2 and P3.

For `module.py`, it looks like I confused the `urls` local variable
for an already imported `werkzeug.urls` and just tried calling
`url_parse` on that. Which doesn't work, because `urls` is a
dict.

Fixes #71076

closes odoo/odoo#71120

X-original-commit: 2abc23e2a9300cddf3370fe0ac83b005318c825f
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>
Signed-off-by: Adrian Torres (adt) <adt@odoo.com>