[ADD] pos_sale_product_configurator: add optional products in the product info popup
With the module sale_product_configurator, you can add optionnal
products on your product to suggest them during the sale process.

We are now displaying them of the info popup of a product to easily find
them in the POS, and we also allow to automatically open this popup for
products having such optionnal products.

closes odoo/odoo#71203

Task-id: 2518884
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>